requirements.txt
pytest-playwright
pydantic
toml
pytest-asyncio
pytest-html
aiohttp